+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 - controller: removed service_name from metrics
 - controller: Add error context to the keycloak-admin-client
+- controller: added the optional claim `nickname` to the login endpoint that will be used as the users `display_name` when set
 
 # 1.0.0-rc.3 (20 July, 2022)